refactor LotrGame into useLotrGameSetup, useCardActions and useArrowActions for clearity
